feat: extract entity processing
