+++ Vesta/Model/Sync/EntityProcessors/EntityProcessor.swift
@@ -1,0 +1,20 @@
+import Foundation
+import OSLog
+import SwiftData
+
+protocol EntityProcessor {
+    var modelContext: ModelContext { get }
+    var logger: Logger { get }
+
+    func process(entities: [[String: Any]], currentUser: User) async throws
+}
+
+class BaseEntityProcessor {
+    let modelContext: ModelContext
+    let logger: Logger
+
+    init(modelContext: ModelContext, logger: Logger) {
+        self.modelContext = modelContext
+        self.logger = logger
+    }
+}
