+++ Vesta/Model/Sync/SyncService.swift
@@ -1,5 +1,6 @@
 import Combine
 import Foundation
+import OSLog
 import SwiftData
 import os
 
@@ -20,6 +21,7 @@ class SyncService: ObservableObject {
     private var todoItemCategories: TodoItemCategoryService
     private var recipes: RecipeService
     private var shoppingItems: ShoppingListItemService
+    private var entityProcessorCoordinator: EntityProcessorCoordinator
 
     private var cancellables = Set<AnyCancellable>()
     private var syncTimer: Timer?
@@ -52,6 +54,18 @@ class SyncService: ObservableObject {
         self.todoItems = todoItems
         self.recipes = recipes
         self.shoppingItems = shoppingItems
+
+        // Initialize the entity processor coordinator
+        self.entityProcessorCoordinator = EntityProcessorCoordinator(
+            modelContext: modelContext,
+            users: users,
+            todoItems: todoItems,
+            todoItemCategories: todoItemCategories,
+            meals: meals,
+            recipes: recipes,
+            shoppingItems: shoppingItems,
+            logger: logger
+        )
     }
 
     /// Start sync operations: initial pull + subscribe to changes + periodic push
@@ -470,28 +484,14 @@ class SyncService: ObservableObject {
                     await MainActor.run {
                         Task {
                             do {
-                                // Process each entity type
-                                for (entityType, entities) in entityData {
-                                    switch entityType {
-                                    case "User":
-                                        try await self.processUserEntities(
-                                            entities, currentUser: currentUser)
-                                    case "TodoItem":
-                                        try await self.processTodoItemEntities(
-                                            entities, currentUser: currentUser)
-                                    case "Recipe":
-                                        try await self.processRecipeEntities(
-                                            entities, currentUser: currentUser)
-                                    case "Meal":
-                                        try await self.processMealEntities(
-                                            entities, currentUser: currentUser)
-                                    case "ShoppingListItem":
-                                        try await self.processShoppingListItemEntities(
-                                            entities, currentUser: currentUser)
-                                    default:
-                                        print("Unknown entity type: \(entityType)")
-                                    }
-                                }
+                                // Map the API entity types to our internal structure
+                                let mappedEntityData = self.mapEntityTypes(entityData)
+
+                                // Process all entities using the coordinator
+                                try await self.entityProcessorCoordinator.processEntities(
+                                    mappedEntityData,
+                                    currentUser: currentUser
+                                )
 
                                 // Save all changes
                                 try self.modelContext.save()
@@ -499,13 +499,13 @@ class SyncService: ObservableObject {
                                 // Complete successfully
                                 promise(.success(()))
                             } catch {
-                                print("Error processing entities: \(error)")
+                                self.logger.error("Error processing entities: \(error)")
                                 promise(.failure(.unknown))
                             }
                         }
                     }
                 } catch {
-                    print("Error in MainActor scheduling: \(error)")
+                    self.logger.error("Error in MainActor scheduling: \(error)")
                     await MainActor.run {
                         promise(.failure(.unknown))
                     }
@@ -514,533 +514,39 @@ class SyncService: ObservableObject {
         }
     }
 
-    @MainActor
-    private func processUserEntities(_ entities: [[String: Any]], currentUser: User) async throws {
-        self.logger.info("Processing \(entities.count) User entities")
-
-        for data in entities {
-            guard let uid = data["uid"] as? String else {
-                self.logger.warning("Skipping User entity without UID")
-                continue
-            }
-
-            // Check if entity exists or create a new one
-            let user: User
-            if let existingUser = try users.fetchUnique(withUID: uid) {
-                user = existingUser
-                self.logger.debug("Found existing User with UID: \(uid)")
-            } else {
-                // For new users, we need just the UID - the rest will be updated
-                user = User(uid: uid)
-                modelContext.insert(user)
-                self.logger.debug("Created new User with UID: \(uid)")
-            }
-
-            // Update properties and members
-            user.update(from: data)
-
-            if let ownerId = data["ownerId"] as? String {
-                if ownerId != user.owner?.uid {
-                    if ownerId == uid {
-                        user.owner = user
-                    } else if let owner = try? users.fetchUnique(withUID: ownerId) {
-                        user.owner = owner
-                    }
-                }
-            } else if user.owner != nil {
-                user.owner = nil
-            }
-
-            // Process friend IDs if available
-            if let friendIds = data["friendIds"] as? [String], !friendIds.isEmpty {
-                self.logger.debug("Processing \(friendIds.count) friends for user \(uid)")
-
-                // Get current friend IDs for comparison
-                let currentFriendIds = Set(user.friends.compactMap { $0.uid })
-                // Find new friend IDs that need to be added
-                let newFriendIds = Set(friendIds).subtracting(currentFriendIds)
-
-                if !newFriendIds.isEmpty {
-                    self.logger.debug("Adding \(newFriendIds.count) new friends to user \(uid)")
-                    // Fetch and add new friends
-                    if let newFriends = try? users.fetchMany(withUIDs: Array(newFriendIds)) {
-                        user.friends.append(contentsOf: newFriends)
-                    }
-
-                    // If new friends couldn't be found in database, create placeholder users
-                    let fetchedIds = Set(user.friends.compactMap { $0.uid })
-                    let missingIds = newFriendIds.subtracting(fetchedIds)
-
-                    for missingId in missingIds {
-                        let newFriend = User(uid: missingId)
-                        newFriend.markAsSynced()
-                        user.friends.append(newFriend)
-                        self.logger.debug(
-                            "Created placeholder user for friend with UID: \(missingId)")
-                    }
-                }
-
-                // Remove friends that are no longer in the friend list
-                user.friends.removeAll { friend in
-                    guard let friendUid = friend.uid else { return false }
-                    let shouldRemove = !friendIds.contains(friendUid)
-                    if shouldRemove {
-                        self.logger.debug("Removing friend with UID \(friendUid) from user \(uid)")
-                    }
-                    return shouldRemove
-                }
-            } else if !user.friends.isEmpty {
-                self.logger.debug("Clearing all friends for user \(uid)")
-                user.friends.removeAll()
-            }
-
-            user.markAsSynced()
-            self.logger.debug("Successfully processed User: \(uid)")
-        }
-    }
-
-    @MainActor
-    private func processTodoItemEntities(_ entities: [[String: Any]], currentUser: User)
-        async throws
+    /// Maps the API entity type names to our internal entity names
+    private func mapEntityTypes(_ entityData: [String: [[String: Any]]]) -> [String: [[String:
+        Any]]]
     {
-        self.logger.info("Processing \(entities.count) TodoItem entities")
+        var mappedData: [String: [[String: Any]]] = [:]
 
-        for data in entities {
-            guard let uid = data["uid"] as? String else {
-                self.logger.warning("Skipping TodoItem entity without UID")
-                continue
+        for (entityType, entities) in entityData {
+            switch entityType {
+            case "User":
+                mappedData["users"] = entities
+            case "TodoItem":
+                mappedData["todoItems"] = entities
+            case "Recipe":
+                mappedData["recipes"] = entities
+            case "Meal":
+                mappedData["meals"] = entities
+            case "ShoppingListItem":
+                mappedData["shoppingListItems"] = entities
+            default:
+                self.logger.warning("Unknown entity type: \(entityType)")
             }
-
-            // Check if entity exists or create a new one
-            let todoItem: TodoItem
-            if let existingTodoItem = try todoItems.fetchUnique(withUID: uid) {
-                todoItem = existingTodoItem
-                self.logger.debug("Found existing TodoItem with UID: \(uid)")
-            } else {
-                guard let title = data["title"] as? String,
-                    let details = data["details"] as? String
-                else {
-                    self.logger.warning(
-                        "Skipping TodoItem without required title or details: \(uid)")
-                    continue
-                }
-
-                todoItem = TodoItem(title: title, details: details, owner: nil)
-                todoItem.uid = uid
-                modelContext.insert(todoItem)
-                self.logger.debug("Created new TodoItem with UID: \(uid), title: \(title)")
-            }
-
-            // Update properties
-            todoItem.update(from: data)
-            self.logger.debug("Updated properties for TodoItem: \(uid)")
-
-            // Update owner if available
-            if let ownerId = data["ownerId"] as? String {
-                if ownerId != todoItem.owner?.uid {
-                    if let owner = try? users.fetchUnique(withUID: ownerId) {
-                        todoItem.owner = owner
-                        self.logger.debug("Set owner \(ownerId) for TodoItem: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find owner with UID \(ownerId) for TodoItem: \(uid)")
-                    }
-                }
-            } else if todoItem.owner != nil {
-                self.logger.debug("Removing owner from TodoItem: \(uid)")
-                todoItem.owner = nil
-            }
-
-            // Process meal reference if available
-            if let mealUID = data["mealId"] as? String {
-                if mealUID != todoItem.meal?.uid {
-                    if let meal = try? meals.fetchUnique(withUID: mealUID) {
-                        todoItem.meal = meal
-                        self.logger.debug("Set meal \(mealUID) for TodoItem: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find meal with UID \(mealUID) for TodoItem: \(uid)")
-                    }
-                }
-            } else if todoItem.meal != nil {
-                self.logger.debug("Removing meal from TodoItem: \(uid)")
-                todoItem.meal = nil
-            }
-
-            // Process shopping list item reference if available
-            if let shoppingListItemUID = data["shoppingListItemId"] as? String {
-                if shoppingListItemUID != todoItem.shoppingListItem?.uid {
-                    if let shoppingListItem = try? shoppingItems.fetchUnique(
-                        withUID: shoppingListItemUID)
-                    {
-                        todoItem.shoppingListItem = shoppingListItem
-                        self.logger.debug(
-                            "Set shopping list item \(shoppingListItemUID) for TodoItem: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find shopping list item with UID \(shoppingListItemUID) for TodoItem: \(uid)"
-                        )
-                    }
-                }
-            } else if todoItem.shoppingListItem != nil {
-                self.logger.debug("Removing shopping list item from TodoItem: \(uid)")
-                todoItem.shoppingListItem = nil
-            }
-
-            // Process category if available
-            if let categoryName = data["categoryName"] as? String {
-                if categoryName != todoItem.category?.name {
-                    todoItem.category = todoItemCategories.fetchOrCreate(named: categoryName)
-                    self.logger.debug("Set category '\(categoryName)' for TodoItem: \(uid)")
-                }
-            } else if todoItem.category != nil {
-                self.logger.debug("Removing category from TodoItem: \(uid)")
-                todoItem.category = nil
-            }
-
-            todoItem.markAsSynced()
-            self.logger.debug("Successfully processed TodoItem: \(uid)")
         }
+
+        return mappedData
     }
 
-    @MainActor
-    private func processRecipeEntities(_ entities: [[String: Any]], currentUser: User) async throws
-    {
-        self.logger.info("Processing \(entities.count) Recipe entities")
+    // Entity processing has been refactored into separate processor classes
 
-        for data in entities {
-            guard let uid = data["uid"] as? String else {
-                self.logger.warning("Skipping Recipe entity without UID")
-                continue
-            }
+    // Entity processing has been refactored into separate processor classes
 
-            // Check if entity exists or create a new one
-            let recipe: Recipe
-            if let existingRecipe = try recipes.fetchUnique(withUID: uid) {
-                recipe = existingRecipe
-                self.logger.debug("Found existing Recipe with UID: \(uid)")
-            } else {
-                guard let title = data["title"] as? String,
-                    let details = data["details"] as? String
-                else {
-                    self.logger.warning("Skipping Recipe without required title or details: \(uid)")
-                    continue
-                }
+    // Entity processing has been refactored into separate processor classes
 
-                recipe = Recipe(
-                    title: title,
-                    details: details,
-                    owner: nil  // Will be updated based on references
-                )
-                recipe.uid = uid
-                recipe.dirty = false  // Fresh from server
-                modelContext.insert(recipe)
-                self.logger.debug("Created new Recipe with UID: \(uid), title: \(title)")
-            }
+    // Entity processing has been refactored into separate processor classes
 
-            // Update properties using the Recipe's update method
-            recipe.update(from: data)
-            self.logger.debug("Updated properties for Recipe: \(uid)")
-
-            // Update owner if available
-            if let ownerId = data["ownerId"] as? String {
-                if ownerId != recipe.owner?.uid {
-                    if let owner = try? users.fetchUnique(withUID: ownerId) {
-                        recipe.owner = owner
-                        self.logger.debug("Set owner \(ownerId) for Recipe: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find owner with UID \(ownerId) for Recipe: \(uid)")
-                    }
-                }
-            } else if recipe.owner != nil {
-                self.logger.debug("Removing owner from Recipe: \(uid)")
-                recipe.owner = nil
-            }
-
-            // Process meal references
-            if let mealIds = data["mealIds"] as? [String], !mealIds.isEmpty {
-                self.logger.debug("Processing \(mealIds.count) meal references for Recipe: \(uid)")
-
-                let currentMealIds = Set(recipe.meals.compactMap { $0.uid })
-                let newMealIds = Set(mealIds).subtracting(currentMealIds)
-
-                if !newMealIds.isEmpty {
-                    if let newMeals = try? meals.fetchMany(withUIDs: Array(newMealIds)) {
-                        recipe.meals.append(contentsOf: newMeals)
-                        self.logger.debug("Added \(newMeals.count) new meals to Recipe: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find any of the \(newMealIds.count) new meals for Recipe: \(uid)"
-                        )
-                    }
-                }
-
-                // Remove meals that are no longer associated with this recipe
-                let initialCount = recipe.meals.count
-                recipe.meals.removeAll { meal in
-                    guard let mealUid = meal.uid else { return false }
-                    return !mealIds.contains(mealUid)
-                }
-                let removedCount = initialCount - recipe.meals.count
-                if removedCount > 0 {
-                    self.logger.debug("Removed \(removedCount) meals from Recipe: \(uid)")
-                }
-            } else if !recipe.meals.isEmpty {
-                self.logger.debug("Clearing all \(recipe.meals.count) meals from Recipe: \(uid)")
-                recipe.meals.removeAll()
-            }
-
-            recipe.markAsSynced()
-            self.logger.debug("Successfully processed Recipe: \(uid)")
-        }
-    }
-
-    @MainActor
-    private func processMealEntities(_ entities: [[String: Any]], currentUser: User) async throws {
-        self.logger.info("Processing \(entities.count) Meal entities")
-
-        for data in entities {
-            guard let uid = data["uid"] as? String else {
-                self.logger.warning("Skipping Meal entity without UID")
-                continue
-            }
-
-            // Check if entity exists or create a new one
-            let meal: Meal
-            if let existingMeal = try meals.fetchUnique(withUID: uid) {
-                meal = existingMeal
-                self.logger.debug("Found existing Meal with UID: \(uid)")
-            } else {
-                guard let scalingFactor = data["scalingFactor"] as? Double,
-                    let mealTypeRaw = data["mealType"] as? String,
-                    let mealType = MealType(rawValue: mealTypeRaw)
-                else {
-                    self.logger.warning("Skipping Meal without required properties: \(uid)")
-                    continue
-                }
-
-                meal = Meal(
-                    scalingFactor: scalingFactor,
-                    todoItem: nil,  // Will be updated later based on references
-                    recipe: nil,  // Will be updated later based on references
-                    mealType: mealType,
-                    owner: nil  // Will be updated later based on references
-                )
-                meal.uid = uid
-                modelContext.insert(meal)
-                self.logger.debug("Created new Meal with UID: \(uid), type: \(mealType.rawValue)")
-            }
-
-            // Update properties
-            meal.update(from: data)
-            self.logger.debug("Updated properties for Meal: \(uid)")
-
-            // Update owner if available
-            if let ownerId = data["ownerId"] as? String {
-                if ownerId != meal.owner?.uid {
-                    if let owner = try? users.fetchUnique(withUID: ownerId) {
-                        meal.owner = owner
-                        self.logger.debug("Set owner \(ownerId) for Meal: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find owner with UID \(ownerId) for Meal: \(uid)")
-                    }
-                }
-            } else if meal.owner != nil {
-                self.logger.debug("Removing owner from Meal: \(uid)")
-                meal.owner = nil
-            }
-
-            // Process recipe reference if available
-            if let recipeUID = data["recipeId"] as? String {
-                if recipeUID != meal.recipe?.uid {
-                    if let recipe = try? recipes.fetchUnique(withUID: recipeUID) {
-                        meal.recipe = recipe
-                        self.logger.debug("Set recipe \(recipeUID) for Meal: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find recipe with UID \(recipeUID) for Meal: \(uid)")
-                    }
-                }
-            } else if meal.recipe != nil {
-                self.logger.debug("Removing recipe from Meal: \(uid)")
-                meal.recipe = nil
-            }
-
-            // Process todoItem reference if available
-            if let todoItemUID = data["todoItemId"] as? String {
-                if todoItemUID != meal.todoItem?.uid {
-                    if let todoItem = try? todoItems.fetchUnique(withUID: todoItemUID) {
-                        meal.todoItem = todoItem
-                        self.logger.debug("Set todoItem \(todoItemUID) for Meal: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find todoItem with UID \(todoItemUID) for Meal: \(uid)")
-                    }
-                }
-            } else if meal.todoItem != nil {
-                self.logger.debug("Removing todoItem from Meal: \(uid)")
-                meal.todoItem = nil
-            }
-
-            // Process shopping list items
-            if let shoppingItemIds = data["shoppingListItemIds"] as? [String],
-                !shoppingItemIds.isEmpty
-            {
-                self.logger.debug(
-                    "Processing \(shoppingItemIds.count) shopping list items for Meal: \(uid)")
-
-                let currentItemIds = Set(meal.shoppingListItems.compactMap { $0.uid })
-                let newItemIds = Set(shoppingItemIds).subtracting(currentItemIds)
-
-                if !newItemIds.isEmpty {
-                    if let newItems = try? shoppingItems.fetchMany(
-                        withUIDs: Array(newItemIds)
-                    ) {
-                        meal.shoppingListItems.append(contentsOf: newItems)
-                        self.logger.debug(
-                            "Added \(newItems.count) new shopping items to Meal: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find any of the \(newItemIds.count) new shopping items for Meal: \(uid)"
-                        )
-                    }
-                }
-
-                // Remove items that are no longer associated with this meal
-                let initialCount = meal.shoppingListItems.count
-                meal.shoppingListItems.removeAll { item in
-                    guard let itemUid = item.uid else { return false }
-                    return !shoppingItemIds.contains(itemUid)
-                }
-                let removedCount = initialCount - meal.shoppingListItems.count
-                if removedCount > 0 {
-                    self.logger.debug("Removed \(removedCount) shopping items from Meal: \(uid)")
-                }
-            } else if !meal.shoppingListItems.isEmpty {
-                self.logger.debug(
-                    "Clearing all \(meal.shoppingListItems.count) shopping items from Meal: \(uid)")
-                meal.shoppingListItems.removeAll()
-            }
-
-            meal.markAsSynced()
-            self.logger.debug("Successfully processed Meal: \(uid)")
-        }
-    }
-
-    @MainActor
-    private func processShoppingListItemEntities(_ entities: [[String: Any]], currentUser: User)
-        async throws
-    {
-        self.logger.info("Processing \(entities.count) ShoppingListItem entities")
-
-        for data in entities {
-            guard let uid = data["uid"] as? String else {
-                self.logger.warning("Skipping ShoppingListItem entity without UID")
-                continue
-            }
-
-            // Check if entity exists or create a new one
-            let shoppingListItem: ShoppingListItem
-            if let existingItem = try shoppingItems.fetchUnique(withUID: uid) {
-                shoppingListItem = existingItem
-                self.logger.debug("Found existing ShoppingListItem with UID: \(uid)")
-            } else {
-                guard let name = data["name"] as? String else {
-                    self.logger.warning("Skipping ShoppingListItem without required name: \(uid)")
-                    continue
-                }
-
-                shoppingListItem = ShoppingListItem(
-                    name: name,
-                    quantity: nil,  // Will be updated from data
-                    unit: nil,  // Will be updated from data
-                    todoItem: nil,  // Will be updated based on references
-                    owner: nil  // Will be updated based on references
-                )
-                shoppingListItem.uid = uid
-                modelContext.insert(shoppingListItem)
-                self.logger.debug("Created new ShoppingListItem with UID: \(uid), name: \(name)")
-            }
-
-            // Update properties
-            shoppingListItem.update(from: data)
-            self.logger.debug("Updated properties for ShoppingListItem: \(uid)")
-
-            // Update owner if available
-            if let ownerId = data["ownerId"] as? String {
-                if ownerId != shoppingListItem.owner?.uid {
-                    if let owner = try? users.fetchUnique(withUID: ownerId) {
-                        shoppingListItem.owner = owner
-                        self.logger.debug("Set owner \(ownerId) for ShoppingListItem: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find owner with UID \(ownerId) for ShoppingListItem: \(uid)")
-                    }
-                }
-            } else if shoppingListItem.owner != nil {
-                self.logger.debug("Removing owner from ShoppingListItem: \(uid)")
-                shoppingListItem.owner = nil
-            }
-
-            // Process todoItem reference if available
-            if let todoItemUID = data["todoItemId"] as? String {
-                if todoItemUID != shoppingListItem.todoItem?.uid {
-                    if let todoItem = try? todoItems.fetchUnique(withUID: todoItemUID) {
-                        shoppingListItem.todoItem = todoItem
-                        self.logger.debug(
-                            "Set todoItem \(todoItemUID) for ShoppingListItem: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find todoItem with UID \(todoItemUID) for ShoppingListItem: \(uid)"
-                        )
-                    }
-                }
-            } else if shoppingListItem.todoItem != nil {
-                self.logger.debug("Removing todoItem from ShoppingListItem: \(uid)")
-                shoppingListItem.todoItem = nil
-            }
-
-            // Process meal references
-            if let mealIds = data["mealIds"] as? [String], !mealIds.isEmpty {
-                self.logger.debug(
-                    "Processing \(mealIds.count) meal references for ShoppingListItem: \(uid)")
-
-                let currentMealIds = Set(shoppingListItem.meals.compactMap { $0.uid })
-                let newMealIds = Set(mealIds).subtracting(currentMealIds)
-
-                if !newMealIds.isEmpty {
-                    if let newMeals = try? meals.fetchMany(withUIDs: Array(newMealIds)) {
-                        shoppingListItem.meals.append(contentsOf: newMeals)
-                        self.logger.debug(
-                            "Added \(newMeals.count) new meals to ShoppingListItem: \(uid)")
-                    } else {
-                        self.logger.debug(
-                            "Failed to find any of the \(newMealIds.count) new meals for ShoppingListItem: \(uid)"
-                        )
-                    }
-                }
-
-                // Remove meals that are no longer associated with this shopping list item
-                let initialCount = shoppingListItem.meals.count
-                shoppingListItem.meals.removeAll { meal in
-                    guard let mealUid = meal.uid else { return false }
-                    return !mealIds.contains(mealUid)
-                }
-                let removedCount = initialCount - shoppingListItem.meals.count
-                if removedCount > 0 {
-                    self.logger.debug("Removed \(removedCount) meals from ShoppingListItem: \(uid)")
-                }
-            } else if !shoppingListItem.meals.isEmpty {
-                self.logger.debug(
-                    "Clearing all \(shoppingListItem.meals.count) meals from ShoppingListItem: \(uid)"
-                )
-                shoppingListItem.meals.removeAll()
-            }
-
-            shoppingListItem.markAsSynced()
-            self.logger.debug("Successfully processed ShoppingListItem: \(uid)")
-        }
-    }
+    // Entity processing has been refactored into separate processor classes
 }
